Extracting read_archive_files to common.py
This removes one of the items that makes verify_deployments.py depend on
deploy.py
